Path Finder Project Using Java
A simple Java Swing application that visualise path finding algorithms (A*, Dijkstra, BFS). Users can interact with the grid to set obstacles, start, and goal points. The algorithms find the shortest path step-by-step, and the program allows users to reset the grid for new puzzles. This project demonstrates GUI development with Java Swing and algorithmic problem-solving in path finding.